+++ src/components/templates/product-table/use-product-actions.tsx
@@ -1,0 +1,85 @@
+import { navigate } from "gatsby"
+import { useAdminDeleteProduct, useAdminUpdateProduct } from "medusa-react"
+import * as React from "react"
+import useImperativeDialog from "../../../hooks/use-imperative-dialog"
+import useToaster from "../../../hooks/use-toaster"
+import { getErrorMessage } from "../../../utils/error-messages"
+import DuplicateIcon from "../../fundamentals/icons/duplicate-icon"
+import EditIcon from "../../fundamentals/icons/edit-icon"
+import PublishIcon from "../../fundamentals/icons/publish-icon"
+import TrashIcon from "../../fundamentals/icons/trash-icon"
+import UnpublishIcon from "../../fundamentals/icons/unpublish-icon"
+import { ActionType } from "../../molecules/actionables"
+import useCopyProduct from "./use-copy-product"
+
+const useProductActions = (product) => {
+  const toaster = useToaster()
+  const dialog = useImperativeDialog()
+  const copyProduct = useCopyProduct()
+  const deleteProduct = useAdminDeleteProduct(product?.id)
+  const updateProduct = useAdminUpdateProduct(product?.id)
+
+  const handleDelete = async () => {
+    const shouldDelete = await dialog({
+      heading: "Delete Product",
+      text: "Are you sure you want to delete this product?",
+    })
+
+    if (shouldDelete) {
+      deleteProduct.mutate()
+    }
+  }
+
+  const getActions = (product): ActionType[] => [
+    {
+      label: "Edit",
+      onClick: () => navigate(`/a/products/${product.id}`),
+      icon: <EditIcon size={20} />,
+    },
+    {
+      label: product.status === "published" ? "Unpublish" : "Publish",
+      onClick: () => {
+        const newStatus = product.status === "published" ? "draft" : "published"
+        updateProduct.mutate(
+          {
+            status: newStatus,
+          },
+          {
+            onSuccess: () => {
+              toaster(
+                `Successfully ${
+                  product.status === "published" ? "unpublished" : "published"
+                } product`,
+                "success"
+              )
+            },
+            onError: (err) => toaster(getErrorMessage(err), "error"),
+          }
+        )
+      },
+      icon:
+        product.status === "published" ? (
+          <UnpublishIcon size={20} />
+        ) : (
+          <PublishIcon size={20} />
+        ),
+    },
+    {
+      label: "Duplicate",
+      onClick: () => copyProduct(product),
+      icon: <DuplicateIcon size={20} />,
+    },
+    {
+      label: "Delete",
+      variant: "danger",
+      onClick: handleDelete,
+      icon: <TrashIcon size={20} />,
+    },
+  ]
+
+  return {
+    getActions,
+  }
+}
+
+export default useProductActions
